Update setup.py
Added -Xpreprocessor flag before the -fopenmp flag to the c compiler, in case the compiler doesn't recognize it even when OpenMP is already installed, for example for macOS 10.14.4 (Mojave). I know the bug can be fixed by just (re)installing gcc through conda but in this way, it will work without going through that process. If the -Xpreprocessor flag doesn't interfere in the non-problematic cases I don't see a problem with adding it.